Instead of waiting on current frame fence, wait on previous fence.
We now signal fences of previous frame when current frame is
updated. This unties the fence life time from overlaylayer
life time and we dont have to wait on current frame fence
like now but ensure that the previous frame kms fence is
signalled before having a new commit. This is needed to
ensure the commit doesn't fail with -EBUSY.

Jira: IAHWC-46
Test: No regressions(Tearing) with apps on Linux and Android.

Signed-off-by: Kalyan Kondapally <kalyan.kondapally@intel.com>